requirements.txt
anthropic>=0.40.0
feedparser>=6.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
rich>=13.0.0
websockets>=12.0
tweepy>=4.14.0
aiohttp>=3.9.0
py-clob-client>=0.18.0
google-genai>=0.8.0
groq>=0.9.0
telethon>=1.36.0
# cache-bust: 2026-04-16
duckduckgo-search>=6.0.0
flask>=3.0.0
